+++ src/components/forms/index.ts
@@ -1,3 +1,2 @@
-export * from "./model";
 export * from "./pipeline";
 export * from "./commons";
