--- src/components/forms/model/ConfigureModelForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConfigureModelForm";
-export * from "./ConfigureModelForm";
