--- src/components/forms/model/ConfigureModelInstanceForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConfigureModelInstanceForm";
-export * from "./ConfigureModelInstanceForm";
