--- src/components/forms/model/CreateModelForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CreateModelForm";
-export * from "./CreateModelForm";
