--- src/components/forms/model/index.ts
@@ -1,21 +1,0 @@
-import ConfigureModelInstanceForm from "./ConfigureModelInstanceForm";
-import type {
-  ConfigureModelInstanceFormProps,
-  ConfigureModelInstanceFormValue,
-} from "./ConfigureModelInstanceForm";
-import CreateModelForm from "./CreateModelForm";
-import type { CreateModelFormValue } from "./CreateModelForm";
-import ConfigureModelForm from "./ConfigureModelForm";
-import type {
-  ConfigureModelFormProps,
-  ConfigureModelFormValue,
-} from "./ConfigureModelForm";
-
-export { ConfigureModelInstanceForm, CreateModelForm, ConfigureModelForm };
-export type {
-  ConfigureModelInstanceFormProps,
-  ConfigureModelInstanceFormValue,
-  CreateModelFormValue,
-  ConfigureModelFormProps,
-  ConfigureModelFormValue,
-};
